REMS-Scheduling-Optimizer-INFORMS/rems-model-web-app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "AchrafAzzaoui",email = "aa270@rice.edu"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-api-python-client (>=2.166.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "gspread-dataframe (>=4.0.0,<5.0.0)",
    "gurobipy (>=12.0.1,<13.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
